regenerate reliable Release files on archive_style simple-subdir
If archive_style is simple-subdir and dynamic_reindex is true, we get in
trouble with 'by-hand' changes in repository. Mini-dinstall looped in
'_gen_release_impl()' over all architecture subdirectories but return by
the first with no changes. Unfortunately 'all' is the first subdir and
here are mostly no changes. All other Release files (for i386,
sparc,...) are broken.